requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic-settings
google-generativeai
python-multipart
pypdf
openpyxl
pandas